Network -> NetworkManager, moved into net.rs for general use
