+++ src/client.rs
@@ -36,8 +36,7 @@ use std::process::exit;
 use std::str::FromStr;
 use std::thread;
 use std::time::Duration;
-use std::collections::VecDeque;
-use net::{RequestAction, ResponseCode, Packet, LineCodec, BroadcastChatMessage};
+use net::{RequestAction, ResponseCode, Packet, LineCodec, BroadcastChatMessage, NetworkManager};
 use tokio_core::reactor::{Core, Timeout};
 use futures::{Future, Sink, Stream, stream};
 use futures::future::ok;
@@ -46,80 +45,6 @@ use futures::sync::mpsc;
 const TICK_INTERVAL:         u64   = 10; // milliseconds
 const CLIENT_VERSION: &str = "0.0.1";
 
-struct NetworkStatistics {
-    packets_tx_failed: u64,
-    packets_tx_success: u64,
-    keep_alive_tx_failed: u64,
-    keep_alive_tx_success: u64,
-}
-
-impl NetworkStatistics {
-    fn new() -> Self {
-        NetworkStatistics {
-            packets_tx_success: 0,
-            packets_tx_failed: 0,
-            keep_alive_tx_failed: 0,
-            keep_alive_tx_success: 0
-        }
-    }
-}
-
-struct Network {
-    statistics:     NetworkStatistics,
-    tx_packets:     VecDeque<Packet>, // Back == Newest, Front == Oldest
-    rx_packets:     VecDeque<Packet>, // Back == Newest, Front == Oldest
-}
-
-impl Network {
-    fn new() -> Self {
-        Network {
-            statistics: NetworkStatistics::new(),
-            tx_packets:  VecDeque::<Packet>::with_capacity(net::PACKET_HISTORY_SIZE),
-            rx_packets:  VecDeque::<Packet>::with_capacity(net::PACKET_HISTORY_SIZE),
-        }
-    }
-
-    fn head_of_tx_packet_queue(&self) -> Option<&Packet> {
-        self.tx_packets.back()
-    }
-
-    fn newest_tx_packet_seq_num(&self) -> Option<u64> {
-        let opt_newest_packet = self.head_of_tx_packet_queue();
-
-        if opt_newest_packet.is_some() {
-            let newest_packet = opt_newest_packet.unwrap();
-            if let Packet::Request{ sequence: newest_sequence, response_ack: _, cookie: _, action: _ } = newest_packet {
-                Some(*newest_sequence)
-            } else { panic!("Found something other than a `Request` packet in the buffer: {:?}", newest_packet) }
-                    // Somehow not a Request packet. Panic during development XXX
-        } else { None }
-                // Queue is empty
-    }
-
-    fn buffer_tx_packet(&mut self, packet: Packet) {
-        let opt_newest_seq_num: Option<u64> = self.newest_tx_packet_seq_num();
-
-        if opt_newest_seq_num.is_none() {
-            self.tx_packets.push_back(packet);
-            return;
-        }
-
-        let newest_sequence = opt_newest_seq_num.unwrap(); // unwrap safe
-
-        if let Packet::Request{ sequence, response_ack: _, cookie: _, action: _ } = packet {
-            // Current packet is newer
-            if newest_sequence < sequence {
-                self.tx_packets.push_back(packet);
-            }
-            // Assumption is that a previous packet with this SQN failed to send,
-            // and this `packet` is up-to-date
-            else if newest_sequence == sequence {
-                self.tx_packets.pop_back();
-                self.tx_packets.push_back(packet);
-            }
-        }
-    }
-}
 
 struct ClientState {
     sequence:         u64,   // sequence number of requests
@@ -131,7 +56,7 @@ struct ClientState {
     cookie:           Option<String>,
     chat_msg_seq_num: u64,
     tick:             usize,
-    network:          Network,
+    network:          NetworkManager,
 }
 
 impl ClientState {
@@ -146,7 +71,7 @@ impl ClientState {
             cookie:          None,
             chat_msg_seq_num: 0,
             tick:            0,
-            network:      Network::new(),
+            network:      NetworkManager::new(),
         }
     }
 
@@ -213,9 +138,9 @@ impl ClientState {
 
                 if send.is_err() {
                     warn!("Could not send UpdateReply{{ {} }} to server", self.chat_msg_seq_num);
-                    self.network.statistics.packets_tx_failed += 1;
+                    self.network.statistics.inc_tx_packets_failed();
                 } else {
-                    self.network.statistics.packets_tx_success += 1;
+                    self.network.statistics.inc_tx_packets_success();
                 }
             }
             Packet::Request{..} => {
@@ -240,9 +165,9 @@ impl ClientState {
 
             if result.is_err() {
                 warn!("Could not send KeepAlive");
-                self.network.statistics.keep_alive_tx_failed += 1;
+                self.network.statistics.inc_tx_keep_alive_failed();
             } else {
-                self.network.statistics.keep_alive_tx_success += 1;
+                self.network.statistics.inc_tx_keep_alive_success();
                 println!("Send KeepAlive yay!")
             }
         }
@@ -299,9 +224,9 @@ impl ClientState {
             let result = (&udp_tx).unbounded_send((addr.clone(), packet));
             if result.is_err() {
                 warn!("Could not send user input cmd to server");
-                self.network.statistics.packets_tx_failed += 1;
+                self.network.statistics.inc_tx_packets_failed();
             } else {
-                self.network.statistics.packets_tx_success += 1;
+                self.network.statistics.inc_tx_packets_success();
             }
         }
     }
